Move the test classes
