+++ vaadin-connect-maven-plugin/src/test/java/com/vaadin/connect/plugin/generator/services/json/JsonTestService.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.connect.plugin.generator.json;
+package com.vaadin.connect.plugin.generator.services.json;
 
 import javax.annotation.security.DenyAll;
 import javax.annotation.security.PermitAll;
