+++ vaadin-connect-maven-plugin/src/test/java/com/vaadin/connect/plugin/generator/services/reservedwordclass/await.java
@@ -13,32 +13,16 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-
-package com.vaadin.connect.plugin.generator.denyall;
-
-import javax.annotation.security.DenyAll;
-import javax.annotation.security.PermitAll;
-import javax.annotation.security.RolesAllowed;
+package com.vaadin.connect.plugin.generator.services.reservedwordclass;
 
 import com.vaadin.connect.VaadinService;
-import com.vaadin.connect.oauth.AnonymousAllowed;
 
+/**
+ * This class is used for OpenApi generator test
+ */
 @VaadinService
-@DenyAll
-public class DenyAllService {
+public class await {
+  public void foo() {
+  }
 
-    public void shouldBeDenied() {
-    }
-
-    @PermitAll
-    public void shouldBeDisplayed1() {
-    }
-
-    @RolesAllowed("test")
-    public void shouldBeDisplayed2() {
-    }
-
-    @AnonymousAllowed
-    public void shouldBeDisplayed3() {
-    }
 }
