+++ vaadin-connect-maven-plugin/src/test/resources/com/vaadin/connect/plugin/generator/services/denyall/expected-DenyAllService.js
