+++ vaadin-connect-maven-plugin/src/test/resources/com/vaadin/connect/plugin/generator/services/json/expected-GeneratorAnonymousAllowedTestClass.js
